cleanup code

+++ public/functions.js
@@ -306,16 +306,37 @@ function go_deeper(){
 
 }
 
+function database_getItem(key){
+    // assert that key is a string 
+    assert (typeof key == "string")
+    if (DATABASE == "browserStorage"){
+        localStorage.getItem(key)
+    }else if (DATABASE == "api"){
+        // get from api
+
+    }
+}
+
+function database_setItem(key,value){
+    // assert that key is a string 
+    assert (typeof key == "string")
+    if (DATABASE == "browserStorage"){
+        localStorage.setItem(key,value)
+    }else if (DATABASE == "api"){
+        // get from api
+
+    }
+}
 
 function getMainGoalsArray(){
-    let arraystring = localStorage.getItem('ga')
+    let arraystring = database_getItem('ga')
     let main_g_arra=  JSON.parse(arraystring)
     return main_g_arra
 }
 
 
 function saveWork(){
-    localStorage.setItem("ga",JSON.stringify(main_goals_array))
+    database_setItem("ga",JSON.stringify(main_goals_array))
     if(user_auth_data.user_unique_key!="" && syncallowed ){
         saveuserDataToDatabase()
     }
@@ -468,17 +489,16 @@ async function getUserGoals(uid){
 
 
 function signOut(){
-
     if (user_auth_data.user_is_logged_in){
         console.log('logging out')
-
-        firebase.auth().signOut().then(res=>console.log(res))
+        if (! prompt("are you sure you want to sign out? this will delete all your data from local/browser storage")){
+            return
+        }
         localStorage.clear()
-        location.reload()
         user_auth_data.user_is_logged_in = false
         user_auth_data = undefined
+        location.reload()
     }
-    // .then(res=>{if res.})
 }
